Unify server layout
